+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 0.1.1
 =====
 
-- Added passwords to default users.
+- Allowed disabling anonymous access to the server.  See the
+  `:allow_anonymous` option on {Ladle::Server#initialize}.
+- Added passwords to default people.
 - Internal: ensured that specs properly close sockets so that the
   suite will pass on Linux.